Refactor tests for _discover_files() into new unittest file.
