+++ tests/lint/unittest_lint.py
@@ -1150,140 +1150,6 @@ def test_recursive_ignore(ignore_parameter: str, ignore_parameter_value: str) ->
     astroid.MANAGER.clear_cache()
 
 
-def mock_isdir(path: str) -> bool:
-    """
-    Mock of os.path.isdir() for the following tests:
-    - test_discover_files_does_not_ignore_similarly_named_package
-    - test_discover_files_does_not_ignore_similarly_named_package_even_if_first_is_ignored
-    - test_discover_files_does_not_ignore_similarly_named_package_and_gets_manage_py
-    """
-    if path == ".":
-        return True
-    raise ValueError("Not expecting an isdir call on %s", path)
-
-
-def mock_isfile(path: str) -> bool:
-    """
-    Mock of os.path.isfile() for the following tests:
-    - test_discover_files_does_not_ignore_similarly_named_package
-    - test_discover_files_does_not_ignore_similarly_named_package_even_if_first_is_ignored
-    - test_discover_files_does_not_ignore_similarly_named_package_and_gets_manage_py
-    """
-    if path == "./__init__.py":
-        return False
-    if path == "./manage.py":
-        return True
-    raise ValueError("Not expecting an isfile call on %s", path)
-
-
-def test_discover_files_does_not_ignore_similarly_named_package(
-    initialized_linter,
-) -> None:
-    """
-    Test to see if we return the expected package/file list even if a shorter named package is processed
-    first and does not match an ignore config value.
-    """
-    with mock.patch("os.walk") as mock_walk:
-        with mock.patch.multiple(
-            "os.path", isdir=mock.DEFAULT, isfile=mock.DEFAULT
-        ) as mock_path:
-            mock_walk.return_value = [
-                (
-                    ".",
-                    ["applications", "applications_api"],
-                    ["pyproject.toml", "manage.py"],
-                ),
-                (
-                    "./applications",
-                    ["tests"],
-                    ["views.py", "models.py", "admin.py", "apps.py", "__init__.py"],
-                ),
-                ("./applications/tests", [], ["test1.py", "test2.py", "__init__.py"]),
-                (
-                    "./applications_api",
-                    ["tests"],
-                    ["views.py", "models.py", "admin.py", "apps.py", "__init__.py"],
-                ),
-                (
-                    "./applications_api/tests",
-                    [],
-                    ["test1.py", "test2.py", "__init__.py"],
-                ),
-            ]
-            mock_path["isdir"].side_effect = mock_isdir
-            mock_path["isfile"].side_effect = mock_isfile
-
-            results = tuple(initialized_linter._discover_files(["."]))
-
-            assert mock_path["isdir"].call_count == 1
-            assert mock_path["isdir"].call_args_list == [mock.call(".")]
-            assert mock_path["isfile"].call_count == 1
-            assert mock_path["isfile"].call_args_list == [
-                mock.call("./__init__.py"),
-            ]
-
-    assert len(results) == 3
-    assert results == ("./manage.py", "./applications", "./applications_api")
-
-
-def test_discover_files_does_not_ignore_similarly_named_package_even_if_first_ignored(
-    initialized_linter,
-) -> None:
-    """
-    Test to see if we return the expected package/file list even if the shorter named package is processed
-    first and matches an ignore config value.
-
-    NOTE: manage.py probably should be ignored.
-    """
-    with mock.patch("os.walk") as mock_walk:
-        with mock.patch.multiple(
-            "os.path", isdir=mock.DEFAULT, isfile=mock.DEFAULT
-        ) as mock_path:
-            initialized_linter.config.ignore = [
-                ".venv",
-                "applications",
-                "node_modules",
-                "manage.py",
-            ]
-            mock_walk.return_value = [
-                (
-                    ".",
-                    ["applications", "applications_api"],
-                    ["pyproject.toml", "manage.py"],
-                ),
-                (
-                    "./applications",
-                    ["tests"],
-                    ["views.py", "models.py", "admin.py", "apps.py", "__init__.py"],
-                ),
-                ("./applications/tests", [], ["test1.py", "test2.py", "__init__.py"]),
-                (
-                    "./applications_api",
-                    ["tests"],
-                    ["views.py", "models.py", "admin.py", "apps.py", "__init__.py"],
-                ),
-                (
-                    "./applications_api/tests",
-                    [],
-                    ["test1.py", "test2.py", "__init__.py"],
-                ),
-            ]
-            mock_path["isdir"].side_effect = mock_isdir
-            mock_path["isfile"].side_effect = mock_isfile
-
-            results = tuple(initialized_linter._discover_files(["."]))
-
-            assert mock_path["isdir"].call_count == 1
-            assert mock_path["isdir"].call_args_list == [mock.call(".")]
-            assert mock_path["isfile"].call_count == 1
-            assert mock_path["isfile"].call_args_list == [
-                mock.call("./__init__.py"),
-            ]
-
-    assert len(results) == 2
-    assert results == ("./manage.py", "./applications_api")
-
-
 def test_source_roots_globbing() -> None:
     run = Run(
         [
